+++ pkg/grpcsrv/middlewares.go
@@ -2,11 +2,13 @@ package grpcsrv
 
 import (
 	"fmt"
+	"runtime/debug"
 
 	grpcmdlw "github.com/grpc-ecosystem/go-grpc-middleware"
 	grpcrecovery "github.com/grpc-ecosystem/go-grpc-middleware/recovery"
 	grpcctxtags "github.com/grpc-ecosystem/go-grpc-middleware/tags"
 	grpcprom "github.com/grpc-ecosystem/go-grpc-prometheus"
+	"github.com/rs/zerolog/log"
 	"google.golang.org/grpc"
 )
 
@@ -15,6 +17,9 @@ func StdUnaryMiddleware(interceptors ...grpc.UnaryServerInterceptor) grpc.Server
 		grpcctxtags.UnaryServerInterceptor(),
 		grpcprom.UnaryServerInterceptor,
 		grpcrecovery.UnaryServerInterceptor(grpcrecovery.WithRecoveryHandler(func(i interface{}) error {
+			log.Error().Str("panic_stack", string(debug.Stack())).
+				Msg("grpc panic")
+
 			return fmt.Errorf("%#v", i)
 		})),
 	}